requirements.txt
streamlit
pillow
exif
collections
pandas
folium
streamlit-folium
os